refactor: use 'number' instead of 'string' in data objects (#8)
The related data is actually a number, so the objects should reflect
this.

### Notes

The value passed as overlay label property is transformed into a String.
This is a current limitation of bpmn-visualization.
If the transform is not done, an error occurs. I will fill an issue in
the bpmn-visualization repository. The label type is string in the API
types but nothing prevents the JavaScript code to pass a number.

In the following stack trace build.js is the mxGraph code!
```
Uncaught TypeError: str.split is not a function
    plainText build.js:20625
    text build.js:20472
    paint build.js:26794
    redrawShape build.js:22994
    redraw build.js:22831
    redraw build.js:26853
    redrawCellOverlays build.js:50100
    visit build.js:1432
    redrawCellOverlays build.js:50074
    redraw build.js:50319
    addCellOverlay build.js:57438
    addOverlays bpmn-visualization.esm.js:5467
    addOverlays bpmn-visualization.esm.js:5465
    addOverlays bpmn-visualization.esm.js:5794
```

Co-authored-by: Souchet Céline <celine.souchet@bonitasoft.com>